Build with waf and dependencies with conan
Integrated the project with conan (conan.io) for dependency
management. Also, used waf build system (https://waf.io)
to easy the build configuration.

The build was not ported exactly (yet?), but may not be needed.
Also, changed the build to generate a library with the server
files (to allow it to be integrated into other projects).